🧪 ADC3A5 fts5: isolate pagination coverage
